refactor: move domain objects to his own namespace
