+++ Doppler.HtmlEditorApi/Controllers/CampaignsController.cs
@@ -9,6 +9,8 @@ using Microsoft.Extensions.Options;
 using Doppler.HtmlEditorApi.ApiModels;
 using Doppler.HtmlEditorApi.Storage;
 using System.Text.Json;
+using Doppler.HtmlEditorApi.Domain;
+using Doppler.HtmlEditorApi.Configuration;
 
 namespace Doppler.HtmlEditorApi.Controllers
 {
