+++ Doppler.HtmlEditorApi/Domain/CampaignState.cs
@@ -2,15 +2,15 @@ using System.Linq;
 using System;
 using System.Collections.Generic;
 
-namespace Doppler.HtmlEditorApi;
+namespace Doppler.HtmlEditorApi.Domain;
 
 public record CampaignState(bool OwnCampaignExists, bool ContentExists, int? EditorType, CampaignStatus? CampaignStatus)
 {
-    private static readonly HashSet<CampaignStatus> WritableStatus = new HashSet<CampaignStatus>
-    (new[]
+    private static readonly HashSet<CampaignStatus> WritableStatus = new HashSet<CampaignStatus>(
+        new[]
         {
-            HtmlEditorApi.CampaignStatus.DRAFT,
-            HtmlEditorApi.CampaignStatus.IN_WINNER_IN_AB_SELECTION_PROCESS
+            Domain.CampaignStatus.DRAFT,
+            Domain.CampaignStatus.IN_WINNER_IN_AB_SELECTION_PROCESS
         });
     public bool IsWritable => CampaignStatus.HasValue && WritableStatus.Contains(CampaignStatus.Value);
 }
