+++ Doppler.HtmlEditorApi/Storage/IFieldsRepository.cs
@@ -1,5 +1,6 @@
 using System.Collections.Generic;
 using System.Threading.Tasks;
+using Doppler.HtmlEditorApi.Domain;
 
 namespace Doppler.HtmlEditorApi.Storage;
